[feat]: Made a separate trait for querying eth bridge related data from storage
