+++ ethereum_bridge/src/storage/mod.rs
@@ -1,4 +1,6 @@
 //! Functionality for accessing the storage subspace
 pub use namada_core::ledger::eth_bridge::storage::bridge_pool;
+pub mod eth_bridge_queries;
 pub mod vote_tallies;
+
 pub use namada_core::ledger::eth_bridge::storage::{wrapped_erc20s, *};
