+++ proof_of_stake/src/pos_queries.rs
@@ -11,11 +11,9 @@ use namada_core::ledger::storage::Storage;
 use namada_core::ledger::{storage, storage_api};
 use namada_core::types::address::Address;
 use namada_core::types::chain::ProposalBytes;
-use namada_core::types::ethereum_events::EthAddress;
 use namada_core::types::key::dkg_session_keys::DkgPublicKey;
 use namada_core::types::storage::{BlockHeight, Epoch};
 use namada_core::types::transaction::EllipticCurve;
-use namada_core::types::vote_extensions::validator_set_update::EthAddrBook;
 use namada_core::types::{key, token};
 use thiserror::Error;
 
@@ -55,17 +53,6 @@ pub enum Error {
 
 /// Result type returned by [`PosQueries`] operations.
 pub type Result<T> = ::std::result::Result<T, Error>;
-
-/// This enum is used as a parameter to
-/// [`PosQueries::can_send_validator_set_update`].
-pub enum SendValsetUpd {
-    /// Check if it is possible to send a validator set update
-    /// vote extension at the current block height.
-    Now,
-    /// Check if it is possible to send a validator set update
-    /// vote extension at the previous block height.
-    AtPrevHeight,
-}
 
 /// Methods used to query blockchain proof-of-stake related state,
 /// such as the currently active set of validators.
@@ -117,10 +104,6 @@ pub trait PosQueries {
         epoch: Option<Epoch>,
     ) -> Result<Address>;
 
-    /// Determines if it is possible to send a validator set update vote
-    /// extension at the provided [`BlockHeight`] in [`SendValsetUpd`].
-    fn can_send_validator_set_update(&self, can_send: SendValsetUpd) -> bool;
-
     /// Check if we are at a given [`BlockHeight`] offset, `height_offset`,
     /// within the current [`Epoch`].
     fn is_deciding_offset_within_epoch(&self, height_offset: u64) -> bool;
@@ -130,29 +113,6 @@ pub trait PosQueries {
 
     /// Retrieves the [`BlockHeight`] that is currently being decided.
     fn get_current_decision_height(&self) -> BlockHeight;
-
-    /// For a given Namada validator, return its corresponding Ethereum bridge
-    /// address.
-    fn get_ethbridge_from_namada_addr(
-        &self,
-        validator: &Address,
-        epoch: Option<Epoch>,
-    ) -> Option<EthAddress>;
-
-    /// For a given Namada validator, return its corresponding Ethereum
-    /// governance address.
-    fn get_ethgov_from_namada_addr(
-        &self,
-        validator: &Address,
-        epoch: Option<Epoch>,
-    ) -> Option<EthAddress>;
-
-    /// Extension of [`Self::get_active_validators`], which additionally returns
-    /// all Ethereum addresses of some validator.
-    fn get_active_eth_addresses<'db>(
-        &'db self,
-        epoch: Option<Epoch>,
-    ) -> Box<dyn Iterator<Item = (EthAddrBook, Address, token::Amount)> + 'db>;
 
     /// Retrieve the `max_proposal_bytes` consensus parameter from storage.
     fn get_max_proposal_bytes(&self) -> ProposalBytes;
@@ -304,28 +264,6 @@ where
             })
     }
 
-    #[cfg(feature = "abcipp")]
-    #[inline]
-    fn can_send_validator_set_update(&self, _can_send: SendValsetUpd) -> bool {
-        // TODO: implement this method for ABCI++; should only be able to send
-        // a validator set update at the second block of an epoch
-        false
-    }
-
-    #[cfg(not(feature = "abcipp"))]
-    #[inline]
-    fn can_send_validator_set_update(&self, can_send: SendValsetUpd) -> bool {
-        if matches!(can_send, SendValsetUpd::AtPrevHeight) {
-            // when checking vote extensions in Prepare
-            // and ProcessProposal, we simply return true
-            true
-        } else {
-            // offset of 1 => are we at the 2nd
-            // block within the epoch?
-            self.is_deciding_offset_within_epoch(1)
-        }
-    }
-
     fn is_deciding_offset_within_epoch(&self, height_offset: u64) -> bool {
         let current_decision_height = self.get_current_decision_height();
 
@@ -359,70 +297,6 @@ where
     #[inline]
     fn get_current_decision_height(&self) -> BlockHeight {
         self.last_height + 1
-    }
-
-    #[inline]
-    fn get_ethbridge_from_namada_addr(
-        &self,
-        validator: &Address,
-        epoch: Option<Epoch>,
-    ) -> Option<EthAddress> {
-        let epoch = epoch.unwrap_or_else(|| self.get_current_epoch().0);
-        self.read_validator_eth_hot_key(validator)
-            .as_ref()
-            .and_then(|epk| epk.get(epoch).and_then(|pk| pk.try_into().ok()))
-    }
-
-    #[inline]
-    fn get_ethgov_from_namada_addr(
-        &self,
-        validator: &Address,
-        epoch: Option<Epoch>,
-    ) -> Option<EthAddress> {
-        let epoch = epoch.unwrap_or_else(|| self.get_current_epoch().0);
-        self.read_validator_eth_cold_key(validator)
-            .as_ref()
-            .and_then(|epk| epk.get(epoch).and_then(|pk| pk.try_into().ok()))
-    }
-
-    #[inline]
-    fn get_active_eth_addresses<'db>(
-        &'db self,
-        epoch: Option<Epoch>,
-    ) -> Box<dyn Iterator<Item = (EthAddrBook, Address, token::Amount)> + 'db>
-    {
-        let epoch = epoch.unwrap_or_else(|| self.get_current_epoch().0);
-        Box::new(self.get_active_validators(Some(epoch)).into_iter().map(
-            move |validator| {
-                let hot_key_addr = self
-                    .get_ethbridge_from_namada_addr(
-                        &validator.address,
-                        Some(epoch),
-                    )
-                    .expect(
-                        "All Namada validators should have an Ethereum bridge \
-                         key",
-                    );
-                let cold_key_addr = self
-                    .get_ethgov_from_namada_addr(
-                        &validator.address,
-                        Some(epoch),
-                    )
-                    .expect(
-                        "All Namada validators should have an Ethereum \
-                         governance key",
-                    );
-                let eth_addr_book = EthAddrBook {
-                    hot_key_addr,
-                    cold_key_addr,
-                };
-                (
-                    eth_addr_book,
-                    validator.address,
-                    validator.bonded_stake.into(),
-                )
-            },
-        ))
     }
 
     fn get_max_proposal_bytes(&self) -> ProposalBytes {
