+++ shared/src/ledger/eth_bridge.rs
@@ -2,3 +2,4 @@
 pub use namada_core::ledger::eth_bridge::storage::wrapped_erc20s;
 pub use namada_core::ledger::eth_bridge::{ADDRESS, INTERNAL_ADDRESS};
 pub use namada_ethereum_bridge::parameters::*;
+pub use namada_ethereum_bridge::storage::eth_bridge_queries::*;
